requirements.txt
openai
pexels-api
moviepy
gtts
google-auth
google-auth-oauthlib
google-api-python-client
requests